Move the inbox rendering inside the inbox package
