+++ internal/cmd/interface.go
@@ -5,10 +5,11 @@ import (
 )
 
 type Inbox interface {
+	inbox.Render
 	ParseInboxPages(int) error
 	Count() int
 	GetMails() []inbox.InboxItem
-	Fetch(inbox.MailKind, int) (inbox.Mail, error)
+	Fetch(inbox.MailKind, int) (inbox.Render, error)
 	Flush() error
 	Delete(int) error
 }
